+++ src/main/java/com/zunf/tankbattleclient/model/message/OutboundMessage.java
@@ -2,19 +2,19 @@ package com.zunf.tankbattleclient.model.message;
 
 public class OutboundMessage {
 
-    private byte type;
+    private int type;
     private byte version;
     private int requestId;
     private byte[] body;
 
-    public OutboundMessage(byte type, byte version, int requestId, byte[] body) {
+    public OutboundMessage(int type, byte version, int requestId, byte[] body) {
         this.type = type;
         this.version = version;
         this.requestId = requestId;
         this.body = body;
     }
 
-    public byte getType() {
+    public int getType() {
         return type;
     }
